+++ src/app/core/models/entity/lear-credential-issuance.ts
@@ -9,7 +9,7 @@ export type IssuanceCredentialType = typeof ISSUANCE_CREDENTIAL_TYPES_ARRAY[numb
 export interface BaseCredentialIssuanceViewModelField {
     key: string, //this is used for form models fields names (FormGroup, FormControl) and also as label for transations; i.e. "mandatee" key is used in "credentialIssuance.mandatee"
     classes?: string; //admits a string of separated classes to customize form styles; i.e.: "classOne classTwo"
-    staticValueGetter?: () => IssuanceStaticViewModel | null; // in case the value must be filled programatically (currently this happens when a field it is 'display: side' or 'pref_side' + asSigner)
+    staticValueGetter?: () => IssuanceStaticViewModel | null; // in case the value must be filled programatically (currently this happens when a field it is 'display: side' or 'pref_side' + onBehalf)
     custom?: { // the Issuance component has some default form templates (text/number input, selector); this field allows for using custom components (i.e. Powers)
       component: ComponentType<BaseIssuanceCustomFormChild<any>>,
       data?: any 
@@ -26,7 +26,7 @@ export interface CredentialIssuanceViewModelControlField extends BaseCredentialI
 
 export interface CredentialIssuanceViewModelGroupField extends BaseCredentialIssuanceViewModelField {
     type: 'group'; // for FormGroup or custom components which include multiple controls
-    display?: 'main' | 'side' | 'pref_side'; // this specifies whether the group should be displayed in the main space or as a side card. 'pref_side' for sections that are only displayed in main when not "asSigner" mode
+    display?: 'main' | 'side' | 'pref_side'; // this specifies whether the group should be displayed in the main space or as a side card. 'pref_side' for sections that are only displayed in main when not "onBehalf" mode
     groupFields: CredentialIssuanceViewModelField[];
 }
 
@@ -62,7 +62,7 @@ export type IssuanceViewModelsTuple = [CredentialIssuanceViewModelSchemaWithId, 
 export interface IssuanceRawCredentialPayload {
   formData: Record<string, any>, 
   staticData: IssuanceStaticViewModel | null,
-  asSigner: boolean
+  onBehalf: boolean
 }
 
 // Power component types
@@ -72,7 +72,7 @@ export interface IssuanceFormPowerSchema{
   //todo: in the future, if there are multiple domains, add a "domain" field (currently there is only "DOME")
   function: string,
   action: string[],
-  isIn2Required: boolean
+  isAdminRequired: boolean
 }
 
 // Key component and service types