+++ src/app/features/credential-issuance/services/issuance-schema-builders/lear-credential-machine-issuance-schema-provider.spec.ts
@@ -138,14 +138,14 @@ describe('LearCredentialMachineIssuanceSchemaProvider', () => {
           {
             action: ['Execute'],
             function: 'Onboarding',
-            isIn2Required: false,
+            isAdminRequired: false,
           },
           {
             action: [
                 "Upload"
             ],
             function: "Certification",
-            isIn2Required: true
+            isAdminRequired: true
             }
         ],
       });