+++ src/app/features/credential-issuance/services/issuance-schema-builders/lear-credential-machine-issuance-schema-provider.ts
@@ -62,7 +62,7 @@ export class LearCredentialMachineIssuanceSchemaProvider implements CredentialIs
           }
         ]
       },
-      // MANDATOR (static when asSigner)
+      // MANDATOR (static when onBehalf)
       {
         key: 'mandator',
         type: 'group',
@@ -110,14 +110,14 @@ export class LearCredentialMachineIssuanceSchemaProvider implements CredentialIs
                 {
                   action: ['Execute'],
                   function: 'Onboarding',
-                  isIn2Required: false
+                  isAdminRequired: false
                 },
                 {
                   action: [
                       "Upload"
                   ],
                   function: "Certification",
-                  isIn2Required: true
+                  isAdminRequired: true
             }
           ]
         }